# internal/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: CalBot API
  version: 0.1.0
  description: API for managing staff authentication and CalBot content.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/staff/menu-options:
    get:
      tags: [Staff menu]
      summary: List every menu option and all associated content, including inactive records
      description: Requires a bearer token belonging to a STAFF account.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All menu options with their prompts, response groups, and response items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StaffMenuOption'
        '401':
          description: Missing or invalid staff token
        '403':
          description: Staff role required
    post:
      tags: [Staff menu]
      summary: Create a menu option
      description: Requires a bearer token belonging to a STAFF account.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMenuOptionRequest'
      responses:
        '201':
          description: Menu option created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuOption'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing or invalid staff token
        '403':
          description: Staff role required
        '409':
          description: Parent option already has response items and cannot receive children
  /api/v1/staff/menu-options/{id}/deactivate:
    patch:
      tags: [Staff menu]
      summary: Deactivate a menu option without deleting it
      description: Requires a bearer token belonging to a STAFF account.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Menu option deactivated
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing or invalid staff token
        '403':
          description: Staff role required
        '404':
          description: Active menu option not found
  /api/v1/staff/menu-options/{id}:
    get:
      tags: [Staff menu]
      summary: Get one menu option without associated prompts or response content
      description: Requires a bearer token belonging to a STAFF account. Active and inactive options can be retrieved.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Menu option table data only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffMenuOptionRow'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing or invalid staff token
        '403':
          description: Staff role required
        '404':
          description: Menu option not found
    delete:
      tags: [Staff menu]
      summary: Permanently delete a menu option and its dependent content
      description: Requires a STAFF bearer token. Deletes child menu options, prompts, response groups, and response items through database cascades. Content feedback is retained with its references set to null.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Menu option and its dependent content deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing or invalid staff token
        '403':
          description: Staff role required
        '404':
          description: Menu option not found
  /api/v1/staff/menu-options/tree:
    get:
      tags: [Staff menu]
      summary: List root menu options for building a tree
      description: Requires a STAFF bearer token. Without parent_id it returns options whose parent_option_id is null. With parent_id it returns direct children, including active and inactive options.
      security:
        - bearerAuth: []
      parameters:
        - name: parent_id
          in: query
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Root menu options, without nested relations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StaffMenuOptionTreeRow'
        '401':
          description: Missing or invalid staff token
        '403':
          description: Staff role required
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    MenuOption:
      type: object
      required: [id, title, sort_order]
      properties:
        id:
          type: integer
          format: int64
        parent_option_id:
          type: integer
          format: int64
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        sort_order:
          type: integer
...
    MenuOptionPrompt:
      type: object
      required: [id, message]
      properties:
        id:
          type: integer
          format: int64
        message:
          type: string
    ResponseGroup:
      type: object
      required: [id, name, items]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/ResponseItem'
    ResponseItem:
      type: object
      required: [id, type, sort_order]
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: [TEXT, LINK, IMAGE, YOUTUBE, VIDEO, DOCUMENT]
        text:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        caption:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        sort_order:
          type: integer
...
    CreateMenuOptionRequest:
      type: object
      required: [parent_option_id, title, sort_order]
      properties:
        parent_option_id:
          type: integer
          format: int64
          nullable: true
        title:
          type: string
        sort_order:
          type: integer
...
    StaffMenuOption:
      allOf:
        - $ref: '#/components/schemas/MenuOption'
        - type: object
          required: [is_active, created_at, updated_at, prompts, response_groups]
          properties:
            is_active:
              type: boolean
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            prompts:
              type: array
              items:
                $ref: '#/components/schemas/StaffMenuOptionPrompt'
            response_groups:
              type: array
              items:
                $ref: '#/components/schemas/StaffResponseGroup'
    StaffMenuOptionRow:
      allOf:
        - $ref: '#/components/schemas/MenuOption'
        - type: object
          required: [is_active, created_at, updated_at]
          properties:
            is_active:
              type: boolean
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    StaffMenuOptionTreeRow:
      allOf:
        - $ref: '#/components/schemas/StaffMenuOptionRow'
        - type: object
          required: [has_response, has_children]
          properties:
            has_response:
              type: boolean
              description: True when the option has at least one stored response item, regardless of active state.
            has_children:
              type: boolean
              description: True when the option has at least one child menu option, regardless of active state.
    StaffMenuOptionPrompt:
      allOf:
        - $ref: '#/components/schemas/MenuOptionPrompt'
        - type: object
          required: [weight, is_active, created_at, updated_at]
          properties:
            weight:
              type: integer
            is_active:
              type: boolean
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    StaffResponseGroup:
      allOf:
        - $ref: '#/components/schemas/ResponseGroup'
        - type: object
          required: [menu_option_id, weight, is_active, created_at, updated_at]
          properties:
            menu_option_id:
              type: integer
              format: int64
            weight:
              type: integer
              minimum: 1
            is_active:
              type: boolean
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            items:
              type: array
              items:
                $ref: '#/components/schemas/StaffResponseItem'
    StaffResponseItem:
      allOf:
        - $ref: '#/components/schemas/ResponseItem'
        - type: object
          required: [is_active, created_at, updated_at]
          properties:
            is_active:
              type: boolean
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string